Player Avatar Configured with a temp camera
The player avatars need to have graphics prefabs that dont have camera for every client,
only the local player should have the camera. For now, temporary testing, I added a camera as child of PlayerAvatar parent, and in ClientAvatarGuidHandler.cs, I activate it if it is localplayer or deactivate it. - In OnNetworkSpawn