RadarSystem problem fixed | CharSelectScene -> MainMenu Scene Reference missing in ClientMainMenuState fixed
RadarSystem problem

A player connects, plays, disconnects, starts a new lobby, and then the radar system is having error, as probably the ClientCharactersCachedInClientMachine has a static List which is not getting deleted on scene unloaded.

CharSelectScene - MainMenu Scene

Solution: The subscription of ClientMainMenuState methods to AuthenticationServiceFacade events were not unsubscribed on destroy .

Hence on next time load of MainMenu Scene, there was becoming duplication of subscriptions.